parseFloat

+++ Scripts/Event/ASIUA;PLANNING!~!~!~.js
@@ -13,10 +13,10 @@ if (tempAsit) {
     var countParcels = 0;
 	for (a in tempAsit) {
         if (!isNaN(tempAsit[a]["Parcel Acreage"])) {
-            parcelAcreage += parseInt(tempAsit[a]["Parcel Acreage"]);
+            parcelAcreage += parseFloat(tempAsit[a]["Parcel Acreage"]);
         }
         if (!isNaN(tempAsit[a]["Revised Acreage"])) {
-            revisedAcreage += parseInt(tempAsit[a]["Revised Acreage"]);
+            revisedAcreage += parseFloat(tempAsit[a]["Revised Acreage"]);
         }
         countParcels++;
 	}//for all rows
